requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.2
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
executing==1.2.0
fastjsonschema==2.16.3
fonttools==4.39.0
fqdn==1.5.1
future==0.18.3
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==8.0.3
jupyter-console==6.6.3
jupyter_core==5.2.0
jupyter-events==0.6.3
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mpmath==1.2.1
mypy==1.1.1
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==3.1.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pylint==2.16.4
pyparsing==3.0.9
pyrsistent==0.19.3
pyspnego==0.8.0
python-constraint==1.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qiskit==0.36.2
qiskit-aer==0.10.4
qiskit-experiments==0.3.1
qiskit-ibmq-provider==0.19.1
qiskit-ignis==0.7.1
qiskit-terra==0.20.2
qtconsole==5.4.0
QtPy==2.3.0
qutip==4.7.1
requests==2.28.2
requests-ntlm==1.2.0
retworkx==0.12.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rustworkx==0.12.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
stevedore==5.0.0
symengine==0.9.2
sympy==1.11.1
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.9.0
tweedledum==1.1.1
typing_extensions==4.5.0
uncertainties==3.1.7
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
wheel==0.38.4
widgetsnbextension==4.0.5
wrapt==1.15.0
